paths shoudl be set up, going to test now

+++ gromosXX/src/check/test_system/test_md_mpi/submit.sh
@@ -1,20 +1,23 @@
 
 . ../gromos_settings.sh
 
-for x in $( ls ${GROMSRC}/testbuild* -d);
+for x in $( ls ${GROMDIR}/testbuild* -d);
 do 
-	echo ${x};
-	binDIR="${x}/bin"
-	dirP=$(basename $x)
-	dirP=${dirP/build_}
+	echo "${x}"
+	if [[ ${x} == *"testbuild"*"mpi"* ]];
+	then
+		echo "TEST ${x}"
+		binDIR="${x}/bin"
+		dirP=$(basename $x)
+		dirP=${dirP/build_}
 
-	echo ${dirP}
-	rm -r ${dirP}
-	cp -r templateMdMpi $dirP
+		echo ${dirP}
+		rm -r ${dirP}
+		cp -r templateMdMpi $dirP
 
-	cd ${dirP}
-	./submit_all.sh ${binDIR}
-	cd ..
-
+		cd ${dirP}
+		./submit_all.sh ${binDIR}
+		cd ..
+	fi;
 done
